Refactoring: Thread rand through calls in test
Instead of accessing a static common rand object, all accesses to
random numbers are now done through arguments to the relevant
functions.

This is an important change, since it paves the way for parallel
execution of tests, which was not possible before due to the shared
single random numebr generator.