Add support for parallel testing
Tests can now be run in parallel using multiple threads.

Note that this puts a lot of high frequency pressure on the memory
allocation system, which with the system allocator on macOS eats up
most of the speed-up. For now, a different memorry allocator, such as
for example mimalloc, needs to be used to get full speed-up.

Added argument -threads that when given a positive number larger than
1 starts that many threads for running the tests. When given 1 it
runs the thread in the main thread, and when given 0 it starts threads
using the number of processing units.

Parallel tests do not work with the log-option, since that uses a
global ostringstream. This combination is checked for and forbidden
when parsing options.